pyproject.toml
[tool.poetry]
name = "testtasks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Django = "^3.2.7"
django-environ = "^0.6.0"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.4.3"
flake8-eradicate = "^1.1.0"
flake8-isort = "^4.0.0"
flake8-print = "^4.0.0"
isort = "^5.9.3"
pytest-django = "^4.4.0"
pytest = "^6.2.5"
pytest-env = "^0.6.2"
pytest-mock = "^3.6.1"
mixer = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
  | migrations
)/

'''
